refactor(services): extract status-page event timeline into shared module
